pyproject.toml
[project]
name = "flywheel"
version = "0.1.0"
description = "Orchestration framework for measurable AI improvement loops"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
    # Batteries dependencies (needed for testing batteries/ code).
    "anyio",
    "claude-agent-sdk",
    "httpx",
    "mcp",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "live_api: requires live Anthropic API access (network + auth); opt in with --run-live-api",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import ordering)
    "UP",     # pyupgrade (Python version idioms)
    "B",      # flake8-bugbear (common pitfalls)
    "SIM",    # flake8-simplify
    "D",      # pydocstyle (docstring conventions)
    "PLC0415",  # no imports inside functions
]
ignore = [
    "SIM108", # ternary operator suggestions (often less readable)
    "D104",   # missing docstring in __init__.py
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]       # no docstring requirements for tests
"tests/test_import.py" = ["PLC0415"]  # import inside function is intentional

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["flywheel"]
